release: prepare v1.23.0
