+++ VERSION
@@ -1,1 +1,1 @@
-1.22.4
+1.23.0
