+++ composer.json
@@ -14,8 +14,7 @@
 		"automattic/jetpack-autoloader": "^5.0.19",
 		"psr/simple-cache": "^1.0 || ^2.0 || ^3.0",
 		"smalot/pdfparser": "^2.12.5",
-		"x-wp/di": "dev-fix/rest-context-plain-permalink-support as 1.9.99",
-		"yahnis-elsts/plugin-update-checker": "^5.6"
+		"x-wp/di": "dev-fix/rest-context-plain-permalink-support as 1.9.99"
 	},
 	"repositories": [
 		{
@@ -33,6 +32,7 @@
 		"squizlabs/php_codesniffer": "^3.7",
 		"szepeviktor/phpstan-wordpress": "^2.0",
 		"wp-coding-standards/wpcs": "^3.0",
+		"yahnis-elsts/plugin-update-checker": "^5.6",
 		"yoast/phpunit-polyfills": "^3.0"
 	},
 	"autoload": {
