+++ includes/Core/SuperdavSiteConnectionService.php
@@ -31,7 +31,6 @@ final class SuperdavSiteConnectionService {
 	private const ACCOUNT_STATUS_ENDPOINT_PATH            = 'site/account';
 	private const ACCOUNT_ACTION_ENDPOINT_PATH            = 'site/account/action';
 	private const ACCOUNT_COUPON_REDEMPTION_ENDPOINT_PATH = 'site/account/redeem-coupon';
-	private const ADVANCED_PLUGIN_METADATA_ENDPOINT       = 'https://sdaiagent.com/?sdai_update_action=get_metadata&sdai_update_slug=superdav-ai-agent-advanced';
 
 	/**
 	 * Maximum number of newest-first credit activity rows retained for display.
@@ -427,78 +426,6 @@ final class SuperdavSiteConnectionService {
 	 */
 	public function has_remote_registration_endpoint(): bool {
 		return '' !== $this->get_registration_endpoint();
-	}
-
-	/** Return the public PUC metadata endpoint for Advanced. */
-	public function get_advanced_plugin_metadata_endpoint(): string {
-		$endpoint = defined( 'SD_AI_AGENT_ADVANCED_PLUGIN_METADATA_ENDPOINT' ) && is_string( SD_AI_AGENT_ADVANCED_PLUGIN_METADATA_ENDPOINT )
-			? SD_AI_AGENT_ADVANCED_PLUGIN_METADATA_ENDPOINT
-			: self::ADVANCED_PLUGIN_METADATA_ENDPOINT;
-		$endpoint = apply_filters( 'sd_ai_agent_advanced_plugin_metadata_endpoint', $endpoint );
-
-		return $this->sanitize_account_url( $endpoint );
-	}
-
-	/**
-	 * Fetch and validate the Advanced package metadata used for installation.
-	 *
-	 * @return array{version:string,download_url:string,package_sha256:string}|WP_Error
-	 */
-	public function request_advanced_plugin_metadata(): array|WP_Error {
-		$endpoint = $this->get_advanced_plugin_metadata_endpoint();
-		if ( '' === $endpoint ) {
-			return new WP_Error( 'sd_ai_agent_advanced_package_unavailable', __( 'SD AI Agent Advanced is temporarily unavailable.', 'superdav-ai-agent' ), array( 'status' => 503 ) );
-		}
-
-		$response = wp_remote_get(
-			$endpoint,
-			array(
-				'timeout'     => 15.0,
-				'redirection' => 0,
-				'headers'     => array( 'Accept' => 'application/json' ),
-			)
-		);
-
-		if ( is_wp_error( $response ) ) {
-			return new WP_Error( 'sd_ai_agent_advanced_package_unavailable', __( 'SD AI Agent Advanced is temporarily unavailable.', 'superdav-ai-agent' ), array( 'status' => 503 ) );
-		}
-
-		$status = (int) wp_remote_retrieve_response_code( $response );
-		if ( $status < 200 || $status >= 300 ) {
-			return new WP_Error( 'sd_ai_agent_advanced_package_unavailable', __( 'SD AI Agent Advanced is temporarily unavailable.', 'superdav-ai-agent' ), array( 'status' => 503 ) );
-		}
-
-		$body           = json_decode( wp_remote_retrieve_body( $response ), true );
-		$download_url   = is_array( $body ) ? $this->sanitize_account_url( $body['download_url'] ?? '' ) : '';
-		$version        = is_array( $body ) && isset( $body['version'] ) && is_string( $body['version'] ) ? $body['version'] : '';
-		$download_query = array();
-		parse_str( (string) wp_parse_url( $download_url, PHP_URL_QUERY ), $download_query );
-		if ( ! is_array( $body )
-			|| 'superdav-ai-agent-advanced' !== ( $body['slug'] ?? null )
-			|| ! isset( $body['version'], $body['download_url'], $body['package_sha256'] )
-			|| ! is_string( $body['version'] )
-			|| ! is_string( $body['download_url'] )
-			|| ! is_string( $body['package_sha256'] )
-			|| 1 !== preg_match( '/^\d+\.\d+\.\d+(?:[-+][0-9A-Za-z.-]+)?$/', $version )
-			|| 1 !== preg_match( '/^[a-f0-9]{64}$/', $body['package_sha256'] )
-			|| '' === $download_url
-			|| wp_parse_url( $endpoint, PHP_URL_SCHEME ) !== wp_parse_url( $download_url, PHP_URL_SCHEME )
-			|| wp_parse_url( $endpoint, PHP_URL_HOST ) !== wp_parse_url( $download_url, PHP_URL_HOST )
-			|| wp_parse_url( $endpoint, PHP_URL_PORT ) !== wp_parse_url( $download_url, PHP_URL_PORT )
-			|| wp_parse_url( $endpoint, PHP_URL_PATH ) !== wp_parse_url( $download_url, PHP_URL_PATH )
-			|| array(
-				'sdai_update_action' => 'download',
-				'sdai_update_slug'   => 'superdav-ai-agent-advanced',
-			) !== $download_query
-		) {
-			return new WP_Error( 'sd_ai_agent_advanced_package_invalid', __( 'The SD AI Agent Advanced package response was invalid.', 'superdav-ai-agent' ), array( 'status' => 502 ) );
-		}
-
-		return array(
-			'version'        => $version,
-			'download_url'   => $download_url,
-			'package_sha256' => $body['package_sha256'],
-		);
 	}
 
 	/**
