+++ package.json
@@ -1,6 +1,6 @@
 {
 	"name": "superdav-ai-agent",
-	"version": "1.22.4",
+	"version": "1.23.0",
 	"private": true,
 	"license": "GPL-2.0-or-later",
 	"packageManager": "pnpm@11.20.0",
